+++ src/index.ts
@@ -1,5 +1,6 @@
 // vim: tabstop=8 softtabstop=0 noexpandtab shiftwidth=8 nosmarttab
 
+export * from './base-params.js';
 export * from './recipe-template.schema.js';
 export * from './recipe.schema.js';
 export * from './request.schema.js';