+++ src/recipe.schema.ts
@@ -1,6 +1,7 @@
 // vim: tabstop=8 softtabstop=0 noexpandtab shiftwidth=8 nosmarttab
 
 import { z } from 'zod/v4';
+import { BaseParams } from './base-params.js';
 
 export namespace RecipeSchema {
 
@@ -29,7 +30,7 @@ export namespace RecipeSchema {
 			.describe("Subresource Integrity (SRI) value"),
 		duration: z.number().min(1).max(86400)
 			.describe("Duration of the image in seconds"),
-		params: z.any()
+		params: BaseParams.optional()
 			.describe("Optional parameters of the image element"),
 	})
 		.describe("HTML image element");
@@ -51,7 +52,7 @@ export namespace RecipeSchema {
 			.describe("Subresource Integrity (SRI) value"),
 		duration: z.number().min(1).max(86400)
 			.describe("Duration of the video in seconds"),
-		params: z.any()
+		params: BaseParams.optional()
 			.describe("Optional parameters of the video element"),
 	})
 		.describe("HTML video element");
@@ -91,13 +92,13 @@ export namespace RecipeSchema {
 			.describe("Subresource Integrity (SRI) value"),
 		duration: z.number().min(1).max(86400)
 			.describe("Duration of the custom element in seconds"),
-		params: z.any()
+		params: BaseParams.optional()
 			.describe("Optional parameters of the custom element"),
 		sources: z.array(z.union([
 			HTMLImageElement.omit({ duration: true }),
 			HTMLVideoElement.omit({ duration: true }),
 			HTMLScriptElement,
-		]))
+		])).optional()
 			.describe("Array of sources, which can be HTMLImageElement, HTMLVideoElement, or HTMLScriptElement"),
 	})
 		.describe("Custom element");
@@ -222,6 +223,14 @@ export namespace RecipeSchema {
 			.describe("Subresource Integrity (SRI) value"),
 		duration: z.number().min(1).max(86400)
 			.describe("Duration of the transition in seconds"),
+		params: BaseParams.optional()
+			.describe("Optional parameters of the transition"),
+		sources: z.array(z.union([
+			HTMLImageElement.omit({ duration: true }),
+			HTMLVideoElement.omit({ duration: true }),
+			HTMLScriptElement,
+		])).optional()
+			.describe("Array of sources, which can be HTMLImageElement, HTMLVideoElement, or HTMLScriptElement"),
 	})
 		.describe("Transition");
 	export type Transition = z.infer<typeof Transition>;